Swapped order of comparison
Swapped the order of comparison when calling operator functions. This is
to allow for using the `contains` operator, as well as making the
ordering more intuitive. This now means that when using `lt` (<),
`gt` (>), etc. then the comparison looks like:

`lt` = `result` < `expected`

So, when using a block that looks like:

```yaml
testinfra.user:
  - name: username
  - uid:
      expected: 1000
      comparison: lt
```
then it will evaluate to `True` if the result is less than 1000.