Fixed issue when calling Command module
When calling the Command module, the results of the call would be
returned directly in the stdout/stderr attributes of the module instance
so that the existing logic path was erroring out. Added additional logic
path that will allow for calling attributes that resolve to values and
not just method or property objects.