Added module for initializing Hashicorp Vault
In order to make it easier to get a Vault server fully operational from
scratch this module will handle initializing and (optionall) unsealing a
freshly installed Vault server. If PGP keys and/or Keybase usernames are
provided then the generated sealing keys will be backed up in the vault
server for easy retrieval in encrypted form.