Added EC2 and external_ip grains to repository
Copied the `ec2_info` and `external_ip` grains modules from the
https://github.com/saltstack/salt-contrib repository so that we can
remove that repo as a dependency in the build step for master since they
are the only two pieces that we use. Also, some of the modules in that
repo have since been merged into salt upstream, which means that the
outdated versions in the contrib repo are overriding the updated and
bugfixed versions in Salt.