Patched version of boto_rds from carbon release
The boto_rds module in the carbon release branch has a bug preventing
use of multi_az as well as forcing definition of optional
parameters. This commit copies the patched version that I have submitted
upstream for inclusion in the final release. Once that patch has been
adopted this change can be reverted.