Added functions to alert on expiring leases and delete expired ones
In order to get better advance warning of expiring leases we need a way to scan the current leases and generate an event
or notification for any that are within a certain time horizon. The `scan_leases` function accepts a lease prefix and a
time horizon argument and generates a salt event for any leases that are expiring within that horizon so that we can
then write reactors to handle those events.

In the event that we want to delete any leases that are within a given time horizon of expiration, we can use the
`clean_expired_leases` function.